requirements.txt
pandas
numpy
pillow
plotly
openpyxl
scikit-learn
python-pptx
